+++ src/main/java/com/project/dorumdorum/domain/chat/application/event/ChatNotificationEventListener.java
@@ -3,8 +3,8 @@ package com.project.dorumdorum.domain.chat.application.event;
 import com.project.dorumdorum.domain.chat.domain.entity.Participant;
 import com.project.dorumdorum.domain.chat.domain.service.ParticipantService;
 import com.project.dorumdorum.domain.chat.notification.SseNotificationSender;
-import com.project.dorumdorum.domain.chat.presence.NotificationChannel;
-import com.project.dorumdorum.domain.chat.presence.PresenceService;
+import com.project.dorumdorum.domain.notification.domain.NotificationChannel;
+import com.project.dorumdorum.domain.presence.domain.service.PresenceService;
 import com.project.dorumdorum.domain.notification.domain.service.NotificationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
