+++ src/main/java/com/project/dorumdorum/domain/notification/domain/NotificationChannel.java
@@ -1,4 +1,4 @@
-package com.project.dorumdorum.domain.chat.presence;
+package com.project.dorumdorum.domain.notification.domain;
 
 public enum NotificationChannel {
     STOMP,
