+++ src/main/java/com/project/dorumdorum/domain/presence/domain/entity/PresenceStatus.java
@@ -1,4 +1,4 @@
-package com.project.dorumdorum.domain.chat.presence;
+package com.project.dorumdorum.domain.presence.domain.entity;
 
 public enum PresenceStatus {
     IN_ROOM,
