+++ src/main/java/com/project/dorumdorum/domain/presence/domain/state/PresenceState.java
@@ -1,6 +1,6 @@
-package com.project.dorumdorum.domain.chat.presence.state;
+package com.project.dorumdorum.domain.presence.domain.state;
 
-import com.project.dorumdorum.domain.chat.presence.NotificationChannel;
+import com.project.dorumdorum.domain.notification.domain.NotificationChannel;
 
 public interface PresenceState {
     NotificationChannel decideMessageChannel(String roomId);
